rename pass 1 to pict/libpict
