+++ examples/evtest.c
@@ -2,10 +2,7 @@
 #include <stdio.h>
 #include <stdint.h>
 #include <stdlib.h>
-#include <string.h>
-
-#include <camlib.h>
-#include <ptp.h>
+#include <libpict.h>
 
 struct PtpRuntime r;
