+++ examples/info.c
@@ -1,9 +1,5 @@
 // Scan filesystem
-#include <stdint.h>
-#include <stdlib.h>
-#include <string.h>
-
-#include <camlib.h>
+#include <libpict.h>
 
 int main() {
 	struct PtpRuntime r;
