+++ examples/live.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 #include <stdint.h>
 #include <stdlib.h>
-#include <camlib.h>
+#include <libpict.h>
 
 int main() {
 	struct PtpRuntime r;
