+++ examples/storage.c
@@ -2,9 +2,7 @@
 #include <stdio.h>
 #include <stdint.h>
 #include <stdlib.h>
-#include <string.h>
-
-#include <camlib.h>
+#include <libpict.h>
 
 int main() {
 	struct PtpRuntime *r = ptp_new(0);
