+++ examples/usb.c
@@ -3,9 +3,7 @@
 #include <stdio.h>
 #include <stdint.h>
 #include <stdlib.h>
-#include <string.h>
-
-#include <camlib.h>
+#include <libpict.h>
 
 int main() {
 	struct PtpRuntime *r = ptp_new(PTP_USB);
