+++ examples/wifi.c
@@ -3,9 +3,7 @@
 #include <stdio.h>
 #include <stdint.h>
 #include <stdlib.h>
-#include <string.h>
-
-#include <camlib.h>
+#include <libpict.h>
 
 int ptp_set_prop_value_16(struct PtpRuntime *r, int code, int value) {
 	struct PtpCommand cmd;
