+++ src/cl_enum.h
@@ -2,8 +2,6 @@
 #ifndef CL_ENUM_H
 #define CL_ENUM_H
 
-#define MAX_ENUM_LENGTH 64
-
 /// @defgroup Enums Enums
 /// @brief API to get string that represents PTP code
 /// @addtogroup Enums
