+++ src/libwpd.c
@@ -5,7 +5,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include <camlib.h>
+#include <libpict.h>
 #include <ptp.h>
 #include <libwpd.h>
